third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/html5-import-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3clink%20rel=%22import%22%20href=%22http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%253cscript%253ealert(/XSS/)%253c/script%253e%22%3e' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
